Extract anticheat mappers into contest.anticheat.mapper.ts
contest.mapper.ts grew past 1100 lines as the anticheat config
mapping accumulated; moving it to its own file isolates the
~440 lines of validation + device policy parsing without breaking
existing import paths.

- New contest.anticheat.mapper.ts owns FIXED_SCREEN_SHARE_RECOVERY_GRACE_MS,
  the 5 ensure* validators, mapAnticheatDevicePolicyDto, and
  mapContestAnticheatConfigDto.
- contest.mapper.ts re-exports mapContestAnticheatConfigDto so
  existing callers (repository, tests) keep working.
- contest.mapper.ts: 1110 -> 676 lines.
